Move menu functionality into a separate file. Step 1 on moving it into separate lib
